1 - Extrair/requirements.txt
patool
termcolor